DSEGOG-299 Move `mock_models.py` to a more generic part of the test structure
